+++ cli_test/nextjs-brackets-test/app/blog/[...slug]/page.tsx
@@ -1,0 +1,3 @@
+export default function BlogPage({ params }: { params: { slug: string[] } }) {
+  return <div>Blog: {params.slug.join('/')}</div>;
+}